Refactor and add themes example
