+++ docs/examples/themes/index.md
@@ -1,0 +1,3 @@
+```python
+-8<-- "themes/app.py"
+```
