+++ pytabulator/themes.py
@@ -1,0 +1,9 @@
+from ._utils import set_theme
+
+
+def tabulator_simple():
+    set_theme("tabulator_simple.min.css")
+
+
+def tabulator_midnight():
+    set_theme("tabulator_midnight.min.css")
